requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
Django==6.0
django-celery==3.1.17
exceptiongroup==1.3.1
getmac==0.9.5
idna==3.11
kombu==5.6.1
mysql-connector-python==9.5.0
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psutil==7.1.3
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
pyodbc==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
vine==5.1.0
waitress==3.0.2
wcwidth==0.2.14
